backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
rq==1.15.1
google-generativeai==0.3.1
slowapi==0.1.9